YTPlayer loop controllable via UI
